Added EPL Sample File, Moved code into logical directories
