+++ helpers/print_functions.py
@@ -1,0 +1,40 @@
+"""
+Span League Ranking Application.
+
+~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~
+<print_functions.py> File for useful reusable print functions for Program Terminal info
+                     and Graphics/Art.
+
+Author:             Lloyd McKeen
+Github Username:    ldmckeen
+Email:              ldmckeen@gmail.com
+Company:            Span
+Date:               January 2023
+~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~
+
+Usage: print_cmd_art()
+
+===========================================================================================
+"""
+
+
+def print_cmd_art():
+    """Print Application Heading and Football Graphic."""
+    print()
+    print('=======================================================')
+    print('~~~=~=~=>>>>   SPAN League Ranking Table   <<<<=~=~=~~~')
+    print('=======================================================')
+    print('====      /|      o__        o__     |   |\\        ====')
+    print('====     /x|     /|          /\\      |   |X\\       ====')
+    print('====    /xx|     / > o        <\\     |   |XX\\      ====')
+    print('====-----------------------------------------------====')
+    print('=======================================================')
+    print()
+
+
+def print_program_end():
+    """Print Program End."""
+    print()
+    print('=======================================================')
+    print('~~~=~=~=>>>>          Program END          <<<<=~=~=~~~')
+    print('=======================================================')
