+++ resources/Pycharm Configs/pycharm config for league ranking python app - cmd param input .png
