+++ resources/Pycharm Configs/pycharm config for league ranking python app - text file input .png
