+++ tests/test_league_ranking.py
@@ -17,11 +17,11 @@ Usage: pytest (to run all tests)
 
 ===========================================================================================
 """
-from league_ranking import calculate_ranking
+from league_ranking import calculate_rankings
 
 
-def test_calculate_ranking():
+def test_calculate_rankings():
     """Test Calculate Ranking Function returns Expected output."""
     test_input = 'Lions 3, Snakes 3 Tarantulas 1, FC Awesome 0 Lions 1, FC Awesome 1 ' \
                  'Tarantulas 3, Snakes 1 Lions 4, Grouches 0'
-    assert calculate_ranking(test_input) == test_input
+    assert calculate_rankings(test_input) == test_input
